Refactored Check In to Display & Logic
